refactor naming scheme
Using directories to classify components/pages/etc. is much more preferable
